requirements.txt
streamlit>=1.32.0
tensorflow>=2.15.0
keras>=2.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
scikit-learn>=1.3.0
Pillow>=10.0.0
pandas>=2.0.0
